+++ lib/drip.dart
@@ -1,4 +1,5 @@
-library drip;
-
+export 'src/async_drip.dart';
+export 'src/computed_drip.dart';
 export 'src/drip_core.dart';
+export 'src/drip_observer.dart';
 export 'src/widgets/widgets.dart';