+++ lib/src/widgets/widgets.dart
@@ -1,3 +1,4 @@
 export 'dripper.dart';
+export 'dripping.dart';
 export 'drop.dart';
 export 'dropper.dart';